% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{AddSampleData}
\alias{AddSampleData}
\title{Add Sample Data to VCF}
\usage{
AddSampleData(vcf, sample_sheet)
}
\arguments{
\item{vcf}{An object of class VCF}

\item{sample_sheet}{a data.frame with columns describing the sample and
population that you may wish to filter on; there must be a column labeled
Sample}
}
\value{
returns the vcf that was passed as an argument with it's colData()
  modified to include the population data
}
\description{
Add population specific, and other sample oriented data to your
  VCF
}
