% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{AssignForeground}
\alias{AssignForeground}
\title{Set Foreground and Background Variants for Enrichment}
\usage{
AssignForeground(variants, enrichmentParams)
}
\arguments{
\item{variants}{an object of class overlapsVCF or overlapsGRanges}

\item{enrichmentParams}{an object of class ldEnrichmentParams or a vector of
logicals equal to the length of variants. In the case of the vector of
logicals; TRUE assigns foreground, FALSE assigns background, and NA drops
the element. Objects of class ldEnrichmentParams are valid for overlapsVCF
objects for filtering into foreground or background based upon LD.}
}
\value{
the variants object passed to the function, annotated into foreground
  / not foreground sets
}
\description{
Assigns foreground and background designations to variants.
}
