% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{CalculateEnrichment}
\alias{CalculateEnrichment}
\title{Calculate Enrichment}
\usage{
CalculateEnrichment(variants, features, feature.type = "biofeatures",
  CI = 0.95, prior = c(a = 0.5, b = 0.5), strict.subset = "guess",
  return.overlaps = FALSE)
}
\arguments{
\item{variants}{a list of two GRanges or VCF objects representing the
foregound and background variants to be compared. In the format list(fg =
object, bg = object).}

\item{features}{A GRangesList of biofeatures or segmentations to calculate
enrichment of the variants against. This field is optional if feature.type
= "biofeatures", and the variants objects have already had SetOverlaps
performed against them.}

\item{feature.type}{a character vector, either "biofeatures" for generic
feature overlaps, or "segmenations" specifying that the features slot is
populated with disjoint genomic segmentations imported by
GetSegmentations()}

\item{CI}{numeric, the width of the credible interval for the calculation of
the beta-binomal test for enrichment}

\item{prior}{numeric vector of length two setting the parameters of the prior
of the beta distribution, in the form, e.g., c(a = 0.5, b = 0.5)}

\item{strict.subset}{either a boolean value, or "guess" indicating if the
foreground is a strict subset of the background, or an independant set.}

\item{return.overlaps}{boolean value indicating if, in addition to the
enrichment data.frame, a GRanges object representing the overlaps of the
variants and features be returned.}
}
\value{
either a data.frame indicating the enrichment statistics of the
  variant set in the features, or a list including the overlaps of the the
  variants and features, in addition to the enrichment data.frame
}
\description{
Calculates the enrichment of foreground variants vs background
  variants against a list of biofeatures.
}
