% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{CalculateLD}
\alias{CalculateLD}
\title{Calculate LD}
\usage{
CalculateLD(vcf, index, population, return = "valid")
}
\arguments{
\item{vcf}{An object of class VCF, optionally processed with SetPopulation()
to allow for calculating LD with a specific population.}

\item{index}{A character vector that contains a SNP that is present in the
vcf}

\item{population}{An optional character vector specifying the population upon
which to filter the samples in the vcf}

\item{return}{A character vector, either 'valid' or 'all'. 'valid' returns
valid SNPs from the vcf only, 'all' attempts to return all variants,
including indels. In the case of 'all', still, only SNPs are used for
calculating LD.}
}
\value{
returns the vcf that was passed as an argument with it's rowRanges()
  modified to include the information relative to the index. Including ref,
  and alt alleles, allele frequencies (in the population queried), distance
  to the index, D prime and R squared.
}
\description{
Calculate the Linkage Disequilibrium between an index SNP and the rest of the
SNPs in the VCF
}
