% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{GetBioFeatures}
\alias{GetBioFeatures}
\title{Import Biofeatures}
\usage{
GetBioFeatures(files, genome)
}
\arguments{
\item{files}{A character vector containing a list of files to}

\item{genome}{A character vector describing the genome build against which
the biofeatures were created, e.g. "hg19", "b37", "hg38", "mm10".}
}
\value{
A GRangesList containing the biofeatures imported.
}
\description{
Imports a set of biofeatures in the form of BED files or derivitive formats
for the purpose of annotating variants
}
