% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{GetVariantsInWindow}
\alias{GetVariantsInWindow}
\title{Get all variants in a specified genomic window}
\usage{
GetVariantsInWindow(file, position, genome = "hg19", type = "vcf")
}
\arguments{
\item{file}{A character vector describing a file to read}

\item{position}{An optional GRanges object defining the coordinates of the
file to import}

\item{genome}{A character vector describing the genome build against which
the variants were mapped e.g. "hg19", "b37", "hg38", "mm10".}

\item{type}{A character vector stating that the filetype is either 'vcf' or
'bed'}
}
\value{
an object of class VCF or GRanges representing the variants
}
\description{
Import variants from VCF or BED files, optionally restricted to a specific window.
}
