% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{PlotEnrichment}
\alias{PlotEnrichment}
\title{Plot Enrichment}
\usage{
PlotEnrichment(variant.enrichment, value = "difference", block1 = NULL,
  block2 = NULL, color.by = NULL, colors = NULL, ncol = 1)
}
\arguments{
\item{variant.enrichment}{a data.frame of enrichment data generated by
CalculateEnrichment()}

\item{value}{one of 'difference' or 'oddsratio', thereby plotting the
difference in beta distributions between foreground and background, or the
oddsratio of the comparison.}

\item{block1}{character string defining the column of variant.enrichment to
be used as a factor by which to separate the enrichment into blocks
(separate panels) in the plot. Default is NULL, in which case there is no
blocking.}

\item{block2}{character string defining the column of variant.enrichment to
be used as a factor by which to separate the enrichment into blocks
(separate panels) in the plot. Default is NULL, in which case there is no
blocking.}

\item{color.by}{optional character string defining the column of
variant.enrichment to be used to color the values. If this column has hex
color values, in the format #ff0000 they will be used.}

\item{colors}{optional character vector defining the colors to use for each
value. Must be the same length as the the number of rows in the
variante.enrichment data.frame.}

\item{ncol}{number of columns to use for facet_wrap if only one block is
defined.}
}
\value{
a ggplot plot object
}
\description{
a flexible plotting tool of the enrichment data.frame generated
  by CalculateEnrichment()
}
