% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{SetOverlaps}
\alias{SetOverlaps}
\title{Assign variants to features}
\usage{
SetOverlaps(variants, features)
}
\arguments{
\item{variants}{a VCF or GRanges object representing variants to annotate}

\item{features}{a GRanges or GRangesList object representing features used to
annotate variants.}
}
\value{
returns the variants object that was passed as an argument, VCF
  objects have overlaps stored in rowRanges(), GRanges have overlaps stored
  in mcols(), ShowOverlaps() can be used to reveal this data.
}
\description{
Assign variants to features
}
