% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{SetPopulation}
\alias{SetPopulation}
\title{Set Population in VCF}
\usage{
SetPopulation(vcf, sample_sheet)
}
\arguments{
\item{vcf}{An object of class VCF}

\item{sample_sheet}{a data.frame with columns describing the sample and
population that you may wish to filter on}
}
\value{
returns the vcf that was passed as an argument with it's colData()
  modified to include the population data
}
\description{
Add population specific data to your VCF
}
