% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{ShowOverlaps}
\alias{ShowOverlaps}
\title{Show the overlaps between variants and features}
\usage{
ShowOverlaps(variants, feature = NULL)
}
\arguments{
\item{variants}{a VCF or GRanges object representing annotated variants}

\item{feature}{a character list of features about which to display overlaps, this is optional, if set as NULL, all features are returned.}
}
\value{
a GRanges list representing the overlap between variants and features.
}
\description{
Show the overlaps between variants and features
}
