% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcivar.R
\name{SplitVcfLd}
\alias{SplitVcfLd}
\title{Split VCF by Linkage Disequilibrium}
\usage{
SplitVcfLd(vcf, ld = list(metric = "R.squared", cutoff = 0.8, maf =
  0.01), strict.subset = TRUE)
}
\arguments{
\item{vcf}{an object of class VCF, particularly one that has been processed
by CalcLD()}

\item{ld}{a list containing three fields. These include: \itemize{ \item
metric, either 'R.squared' or 'D.prime' \item cutoff, numeric between 0 and
1, the value of the metric upon which to perform the cutoff \item maf,
numeric between 0 and 0.5, the minor allele frequency considered to be
included in both foreground and background }}

\item{strict.subset}{boolean indicating if one wishes the foreground  to be a
strict subset of the background, or an independant set.}
}
\value{
a list of length 2 containing in the 'fg' slot the VCF object of the
  foreground SNPs, and the bg slot containing the VCF object of the
  background SNPs
}
\description{
Splits a VCF file into foreground and background variant sets based upon a
Linkage Disequilibrium cutoff
}
